Completed autocomplete search feature.
