+++ app/build.gradle
@@ -6,6 +6,7 @@ plugins {
 android {
     compileSdk 32
 
+    useLibrary 'org.apache.http.legacy'
     defaultConfig {
         applicationId "com.example.wheretonext"
         minSdk 29
@@ -34,6 +35,7 @@ dependencies {
     implementation 'com.google.android.material:material:1.6.1'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
     implementation 'androidx.swiperefreshlayout:swiperefreshlayout:1.1.0'
+    implementation 'com.android.volley:volley:1.2.0'
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
