+++ app/src/main/res/drawable/white_border.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item>
+        <shape android:shape="rectangle">
+            <solid android:color="#FFF"/>
+            <corners android:radius="5dp"/>
+        </shape>
+    </item>
+</layer-list>
